Changes to UI & Code
Changes as per review
